app/requirements.txt
auth0-python==4.4.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
Flask==2.3.2
Flask-OAuthlib==0.9.6
Flask-PyMongo==2.3.0
flask-talisman==1.0.0
Flask-WTF==1.2.1
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==2.1.0
packaging==24.0
pandas==2.2.2
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.1
requests-oauthlib==1.1.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==2.3.7
WTForms==3.1.2
